+++ JUMO.UI/ViewModels/PlaylistViewModel.cs
@@ -27,7 +27,7 @@ namespace JUMO.UI
         public int Denominator => Song.Current.Denominator;
         public int TimeResolution => Song.Current.TimeResolution;
 
-        public int ZoomFactor { get; set; } = 4;
+        public double ZoomFactor { get; set; } = 4.0;
 
         public int ZoomPercent
         {
@@ -35,7 +35,7 @@ namespace JUMO.UI
             private set
             {
                 _zoomPercent = Math.Max(ZOOM_PERCENT_MIN, Math.Min(value, ZOOM_PERCENT_MAX));
-                ZoomFactor = (int)(ZOOM_BASE * _zoomPercent / 100.0);
+                ZoomFactor = ZOOM_BASE * _zoomPercent / 100.0;
 
                 OnPropertyChanged(nameof(ZoomPercent));
                 OnPropertyChanged(nameof(ZoomFactor));
